requirements.txt
ncatbot == 3.8.5
openai == 1.60.1
jmcomic
configparser == 7.2.0
img2pdf == 0.6.0
requests~=2.32.3
pyyaml
pikepdf
json_repair
Pillow==11.0.0
imageio==2.36.0
imageio-ffmpeg==0.5.1
flask>=3.0.0
flask-socketio>=5.3.0
python-socketio>=5.10.0
eventlet>=0.34.0
apscheduler
feedparser
